Changed short int to int

+++ Transport.h
@@ -10,5 +10,5 @@ class Transport
 		void InCommon(ifstream& ifst);
 		void OutCommon(ofstream& ofst);
 	protected:
-		short int mData;
+		int mData;
 };
